Add devcontainer definition
devcontainers provide a reproducible development environment. It allow the repo to define an easy-to-use environment for VS Code, GitHub codespaces, and other supporting IDEs.

For example, if a user has the `ms-vscode-remote.remote-containers` extension in VS Code, they can simply provide this repo's URL and VS Code will create a docker-based development environment and put the user's VS Code window into the container.